Bugfix and tune (#24)
Fixed a bug that could occur when setting up a new board to a fresh FEN position.
Negated the effect of improvement on the Futility margin.

ELO   | 3.04 +- 3.36 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 2.97 (-2.94, 2.94) [-3.00, 1.00]
Games | N: 15062 W: 2831 L: 2699 D: 9532

* Bugfix and tune
* Updated version number
Bench 825191

+++ src/board.h
@@ -128,7 +128,7 @@ public:
   Move get_last_move() const { return move_history.back(); }
   BitBoard PlayerBitBoardControl(Color color, BitBoard all_pieces) const;
   bool MoveInListCanRepeat(const std::vector<Move> moves);
-  int CountRepetitions(size_t min_ply = 0) const;
+  int32_t CountRepetitions(int32_t min_ply = 0) const;
 
 private:
   template<int Quiescent, int MoveGenerationType>
